refactor: ParameterlessRndFactoryのパッケージ構造を変更
