+++ src/matsu/num/statistics/random/ArcsineRnd.java
@@ -11,6 +11,7 @@
 package matsu.num.statistics.random;
 
 import matsu.num.statistics.random.accomp.FloatingRandomGenerator;
+import matsu.num.statistics.random.accomp.ParameterlessRndFactory;
 
 /**
  * <p>
