+++ src/matsu/num/statistics/random/StaticBetaRnd.java
@@ -9,6 +9,8 @@
  */
 package matsu.num.statistics.random;
 
+import matsu.num.statistics.random.accomp.ParameterlessRndFactory;
+
 /**
  * <p>
  * 乱数生成のたびにパラメータを指定するベータ分布乱数発生器
