+++ src/matsu/num/statistics/random/accomp/ParameterlessRndFactory.java
@@ -6,9 +6,12 @@
  */
 
 /*
- * 2025.6.7
+ * 2025.7.4
  */
-package matsu.num.statistics.random;
+package matsu.num.statistics.random.accomp;
+
+import matsu.num.statistics.random.Rnd;
+import matsu.num.statistics.random.RndFactory;
 
 /**
  * パラメータを持たない乱数発生器のファクトリを表現する.
@@ -20,7 +23,6 @@ package matsu.num.statistics.random;
  * <p>
  * <u><i>
  * 注意: <br>
- * このインターフェースは将来にわたって維持されるとは限らず,
  * モジュール外ではこのインターフェースを型として依存してはいけない. <br>
  * 公開されたサブインターフェースが存在する場合, 型はサブインターフェースにより扱われなければならない.
  * </i></u>
