+++ src/matsu/num/statistics/random/base/LazyParameterlessRndFactory.java
@@ -13,8 +13,8 @@ package matsu.num.statistics.random.base;
 import java.util.Objects;
 import java.util.function.Supplier;
 
-import matsu.num.statistics.random.ParameterlessRndFactory;
 import matsu.num.statistics.random.Rnd;
+import matsu.num.statistics.random.accomp.ParameterlessRndFactory;
 
 /**
  * <p>
